Add _lp_username() to return username
Make the return codes to have a distinct code for when the username is
disabled (LP_USER_ALWAYS=-1) vs just not shown (logged in user).

Add valid value for LP_USER_ALWAYS: -1, which disables username display.

This prompted a hack in the Powerline theme to get the possibility of no
username to work, so I need to make a more elegant seperator solution.