Add manual page generation to Sphinx docs
Add mappings for config, functions, and theme to manual page sections.
Install does not have a mapping, mostly because to get the manual pages
you already need to have installed Liquidprompt. Though that does leave
out how to install Liquidprompt into the shell. But maybe it is
something that packages should include instructions for.

Reorder where the TOC comes int he theme document, to make the intro
come first in the manual page.