Fixes compatibility with bash-preexec
[bash-preexec](https://github.com/rcaloras/bash-preexec) "provides
preexec and precmd hook functions for Bash in the style of Zsh".

Unfortunately, it interferes with `trap DEBUG` and `PROMPT_COMMAND`
behavior used by `LP_ENABLE_ERROR` and `LP_ENABLE_RUNTIME`.

Fortunately, since liquidprompt already supports add-zsh-hook behavior,
we can use those hooks as guides to where we should add hooks for
bash-preexec.

These hooks get set if __bp_imported is "defined"; this is set by
bash.

Note on tests/test_bash_preexec.sh: test functions' contents are wrapped
in  `( ... )` subshells so that the sourcing and installation of
bash-preexec is confined to that test and does not pollute others.