Update and fix documentation
The documentation libraries were out of date.
Fix the Ubuntu version badge, since the version used is no longer
supported upstream, and therefore has no badge. 22.04 isn't actually
released yet, but already exists and has a badge, and will exist for a
long time as a LTS.
Fix a typo.