Fix Ruby venv docs and add color option
When originally added, the Ruby venv source used the same color variable
option as the Python venv. But when adding the Node.js source, I
realized that it made sense to have its own option, and just default to
the Python venv color instead.

Also fix the version added tags for the Ruby docs.